Make osbs.utils a subpackage
To match ongoing work in atomic-reactor, make osbs.utils a subpackage.
For now, we just have everything defined in __init__.py at the toplevel
scope.

tests/test_utils.py had one utility class along with tests of utils.py -
move that to a 'tests/util.py'.
